move dns sandbox test to separate file
