+++ pkg/networkservice/chains/nsmgr/single_test.go
@@ -36,118 +36,10 @@ import (
 	"github.com/networkservicemesh/sdk/pkg/networkservice/chains/client"
 	"github.com/networkservicemesh/sdk/pkg/networkservice/chains/nsmgr"
 	"github.com/networkservicemesh/sdk/pkg/networkservice/common/excludedprefixes"
-	"github.com/networkservicemesh/sdk/pkg/networkservice/connectioncontext/dnscontext"
 	"github.com/networkservicemesh/sdk/pkg/networkservice/ipam/point2pointipam"
 	"github.com/networkservicemesh/sdk/pkg/tools/clientinfo"
-	"github.com/networkservicemesh/sdk/pkg/tools/dnsutils"
-	"github.com/networkservicemesh/sdk/pkg/tools/dnsutils/cache"
-	"github.com/networkservicemesh/sdk/pkg/tools/dnsutils/dnsconfigs"
-	"github.com/networkservicemesh/sdk/pkg/tools/dnsutils/fanout"
-	"github.com/networkservicemesh/sdk/pkg/tools/dnsutils/memory"
-	"github.com/networkservicemesh/sdk/pkg/tools/dnsutils/next"
-	"github.com/networkservicemesh/sdk/pkg/tools/dnsutils/noloop"
-	"github.com/networkservicemesh/sdk/pkg/tools/dnsutils/norecursion"
-	"github.com/networkservicemesh/sdk/pkg/tools/dnsutils/searches"
 	"github.com/networkservicemesh/sdk/pkg/tools/sandbox"
 )
-
-func requireIPv4Lookup(ctx context.Context, t *testing.T, r *net.Resolver, host, expected string) {
-	addrs, err := r.LookupIP(ctx, "ip4", host)
-	require.NoError(t, err)
-	require.Len(t, addrs, 1)
-	require.Equal(t, expected, addrs[0].String())
-}
-
-func Test_DNSUsecase(t *testing.T) {
-	t.Cleanup(func() { goleak.VerifyNone(t) })
-
-	ctx, cancel := context.WithTimeout(context.Background(), time.Second*200)
-	defer cancel()
-
-	domain := sandbox.NewBuilder(ctx, t).
-		SetNodesCount(1).
-		SetNSMgrProxySupplier(nil).
-		SetRegistryProxySupplier(nil).
-		Build()
-
-	nsRegistryClient := domain.NewNSRegistryClient(ctx, sandbox.GenerateTestToken)
-
-	nsReg, err := nsRegistryClient.Register(ctx, defaultRegistryService(t.Name()))
-	require.NoError(t, err)
-
-	nseReg := defaultRegistryEndpoint(nsReg.Name)
-
-	nse := domain.Nodes[0].NewEndpoint(ctx, nseReg, sandbox.GenerateTestToken)
-
-	dnsConfigsMap := new(dnsconfigs.Map)
-	nsc := domain.Nodes[0].NewClient(ctx, sandbox.GenerateTestToken, client.WithAdditionalFunctionality(dnscontext.NewClient(
-		dnscontext.WithChainContext(ctx),
-		dnscontext.WithDNSConfigsMap(dnsConfigsMap),
-	)))
-
-	dnsConfigs := []*networkservice.DNSConfig{
-		{
-			DnsServerIps:  []string{"127.0.0.1:40053"},
-			SearchDomains: []string{"com"},
-		},
-	}
-
-	// DNS server on nse side
-	dnsRecords := new(memory.Map)
-	dnsRecords.Store("my.domain.", []net.IP{net.ParseIP("4.4.4.4")})
-	dnsRecords.Store("my.domain.com.", []net.IP{net.ParseIP("5.5.5.5")})
-	dnsutils.ListenAndServe(ctx, memory.NewDNSHandler(dnsRecords), ":40053")
-
-	// DNS server on nsc side
-	clientDNSHandler := next.NewDNSHandler(
-		dnsconfigs.NewDNSHandler(dnsConfigsMap),
-		searches.NewDNSHandler(),
-		noloop.NewDNSHandler(),
-		norecursion.NewDNSHandler(),
-		cache.NewDNSHandler(),
-		fanout.NewDNSHandler(fanout.WithTimeout(time.Second)),
-	)
-	dnsutils.ListenAndServe(ctx, clientDNSHandler, ":50053")
-
-	request := &networkservice.NetworkServiceRequest{
-		MechanismPreferences: []*networkservice.Mechanism{
-			{Cls: cls.LOCAL, Type: kernelmech.MECHANISM},
-		},
-		Connection: &networkservice.Connection{
-			Id:             "1",
-			NetworkService: nsReg.Name,
-			Context: &networkservice.ConnectionContext{
-				DnsContext: &networkservice.DNSContext{
-					Configs: dnsConfigs,
-				},
-			},
-			Labels: make(map[string]string),
-		},
-	}
-
-	resolver := net.Resolver{
-		PreferGo: true,
-		Dial: func(ctx context.Context, network, address string) (net.Conn, error) {
-			var dialer net.Dialer
-			return dialer.DialContext(ctx, network, "127.0.0.1:50053")
-		},
-	}
-
-	_, err = resolver.LookupIP(ctx, "ip4", "my.domain")
-	require.Error(t, err)
-
-	conn, err := nsc.Request(ctx, request)
-	require.NoError(t, err)
-
-	requireIPv4Lookup(ctx, t, &resolver, "my.domain", "4.4.4.4")
-	requireIPv4Lookup(ctx, t, &resolver, "my.domain.com", "5.5.5.5")
-
-	_, err = nsc.Close(ctx, conn)
-	require.NoError(t, err)
-
-	_, err = nse.Unregister(ctx, nseReg)
-	require.NoError(t, err)
-}
 
 func Test_AwareNSEs(t *testing.T) {
 	t.Cleanup(func() { goleak.VerifyNone(t) })
